Move surrogate classes to separate folder
