+++ src/NewRemoting/Surrogates/IpAddressSurrogate.cs
@@ -7,7 +7,7 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace NewRemoting
+namespace NewRemoting.Surrogates
 {
 	internal class IpAddressSurrogate : JsonConverter<IPAddress>
 	{
