+++ src/NewRemoting/Surrogates/ProxySurrogate.cs
@@ -6,7 +6,7 @@ using System.Text.Json.Serialization;
 using System.Text.Json;
 using System.Threading.Tasks;
 
-namespace NewRemoting
+namespace NewRemoting.Surrogates
 {
 	internal sealed class ProxySurrogate : JsonConverter<object>
 	{
